Move mapUtil module to js/lib
