+++ src/nyc_trees/js/src/EventMap.js
@@ -7,7 +7,7 @@
 var $ = require('jquery'),
     L = require('leaflet'),
     mapModule = require('./map'),
-    zoom = require('./mapUtil').ZOOM;
+    zoom = require('./lib/mapUtil').ZOOM;
 
 var EventMap = function (options) {
     var map = mapModule.create(options),
