+++ src/nyc_trees/js/src/lib/mapUtil.js
