+++ src/nyc_trees/js/src/reservationPage.js
@@ -4,8 +4,8 @@ var $ = require('jquery'),
     L = require('leaflet'),
     toastr = require('toastr'),
     mapModule = require('./map'),
-    mapUtil = require('./mapUtil'),
-    zoom = require('./mapUtil').ZOOM,
+    mapUtil = require('./lib/mapUtil'),
+    zoom = require('./lib/mapUtil').ZOOM,
     SelectableBlockfaceLayer = require('./lib/SelectableBlockfaceLayer'),
 
     dom = {
